backend/requirements.txt
cryptography==43.0.3
Faker==30.8.2
Flask==3.0.3
Flask_Bcrypt==1.0.1
flask_cors==5.0.0
psycopg2_binary==2.9.9
PyJWT==2.9.0
PyJWT==2.9.0
python-dotenv==1.0.1
